Fix DLAGENTS for automatic download
Prevent existing DLAGENTS arrays (configured in /etc/makepkg.conf and
~/.makepkg.conf) from being overwritten by the DLAGENTS array in the
PKGBUILD.

Without the plus (+) sign, the DLAGENTS array in the PKGBUILD would
overwrite and replace the existing DLAGENTS array, effectively disabling
all download handlers for this PKGBUILD.

The proper way to handle is to append (+=) to the array so that any
existing handler for hib::/ remains in effect, and the fallback handler
in the PKGBUILD only applies with low priority, i.e. if no other
hib::/ handler is configured in /etc/makepkg.conf nor ~/.makepkg.conf.